admins: new admin group to manage bulk jobs on Phabricator
This creates a new shell admin group on Phabricator servers
that will allow members to manage bulk jobs and configure
them to execute silently.

The use case is to move tasks from one workboard to
another in batches without creating notification spam.

The sudo privilege line is a subset of the existing
privileges for phabricator-admins.

More details as described by Max Binder on the ticket below.

Creating a user for him and adding him to the group will be
a separate change.

Bug: T251349
Change-Id: I683671dd328022a6ba60b3f567a0dc5fc642417e