ATS: stop unconditionally caching 404s
Our Lua implementation of TTL caps for 404 responses is incorrect. Not
only does it set the TTL to 10 minutes even if the origin specified a
lower TTL, it also caches CC:private responses, which is definitely
wrong. We do not currently have any way to know in Lua if the response
is cacheable and what the TTL is other than parsing Cache-Control.

Bug: T250815
Change-Id: Ibaa01d68dac1609f99bba6ae95de25a4c6bae894