aptrepo: kubeadm-k8s: create versioned components
The reprepro software doesn't support keeping multiple versions of the same
packages in the same components. For migration purposes we need to keep kubeadm
and friends in version 1.15 and 1.16 at the same time somewhere in the repo.

This patch reworks the thirdparty/kubeadm-k8s component and splits it into 2:

 * thirdparty/kubeadm-k8s-1-15 -- kubeadm and friends in the 1.15 branch
 * thirdparty/kubeadm-k8s-1-16 -- kubeadm and friends in the 1.16 branch

While at it, update clients that used the old combined repo component to use
the v1.15 one, which is the equivalent of the old one.

Bug: T250866
Change-Id: I9461427712bc82cd1f7cb1806430c4ae7c759bba
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>