mtail: update varnishrls compatibility with rc35
mtail 3.0.0-rc35 does variable type detection during the program parsing phase.  in kind, $maxage
is being typed as an integer.

at runtime, encountering a log entry with no maxage match makes $maxage an empty string and
errors attempting to coerce it into an int (using strconv.ParseInt) and ultimately drops it.

this is a known bug: https://github.com/google/mtail/issues/267

this cs mangles the match a bit so that mtail will not treat it like an integer, but a string

Bug: T251466
Change-Id: Ia764570b2d81c861e9fc7b5d52f56354006be6ec