Enable base::service_auto_restart for apache2 on xhgui
Automatically restart Apache if underlying runtime library dependencies
have been upgraded. The restart is splayed across each day (on Mo-Fr).

Tested on xhgui2001 after applying the latest openldap security update.

This is limited to Buster/Stretch since the systemd support for Apache in jessie
is lacking, it gets generated from the sysvinit script, which makes
it impossible to spot the dependencies on the MPMs.

Bug: T135991
Change-Id: Ifb67cf072dd3c1cf75fcb5d6aa2624c55f7b48bd