jupyterhub: allow notebooks to write under /run/user
What:
Allow the systemd units for jupyterhub notebooks to write
under /run/user.

Why:
We are moving the kerberos credential cache under /run/user,
so the related systemd units will need to be able to write under
it to create kerberos credential tickets. This will avoid to have
duplicates if the user kinits via ssh shell or jupyter terminal.

Tried to limit the rw dirs to /run/user/{USERID} but it seems that
systemd spawner expands USERID in the wrong way (in my tests, it was
1). Since the dirs under /run/user have already the right permissions
and every user unit is ran as its corresponding posix user, it shouldn't
be an issue to use /run/user in my opinion.

Tested on stat1005, I can kinit on the ssh shell and see the credentials
via the jupyter terminal via klist.

Change-Id: I5e419074c49289d4827f208d3e8456e27fee3ba3