mariadb: Enable read_only check to page on primary masters
Enable paging of the icinga check that makes sure that core
masters (metadata, external storage) and parsercaches (on
eqiad now, on codfw when that is the primary datacenters) page
if they go read only- for example, if they go down & restart.

This was scheduled to be enabled after an incident happened,
and in fact the check was shortly enabled, but never was set
to paging after testing time. Do it now.

Refactor a bit the boolean flags on core for readibility and
maintainability.

Bug: T172489
Change-Id: I0e69913d8d10ea527550b9f063e623083d4fa7a6