deployment_server: in cloud we need more space in /srv
This profile mounts a second local disk at /srv to allocate
extra space.

This is needed for a deployment_server to work in cloud projects.

The current workaround for that is that instances have a second
role(sic!) applied and we should avoid that.

It should just work with the regular role like in prod.

Change-Id: I618e37ec8a7af56deef2019856830eee0a13e718