redis: Allow override of instance config
Why:
For the migration of the ORES redis database to the main misc redis
databases, we need to be able to override the redis configuration to
alter the database size to a smaller one and most importantly disable
persistence

What:
Add an $instance_overrides parameter that is a Hash that allows pass
specific directives per instance. Merge that override in the
configuration generated before passing it to the underlying profile.
Switch to using an each() statement to achieve the above.

Change-Id: Ib8076e5f5ca8c946fe84486efe953ecf53f3db80