logstash: Set tcp input type to class $title
logstash::input::udp was setting $type = $title, but ::tcp was not.
The downstream logstash filters all assume that $type = $title, end
result is that input coming over the single definition using
logstash::input::tcp  was flagged as tcp instead of json_lines, not
filter was setup for the tcp type, and messages were dropped on the
floor rather than being indexed into elasticsearch.

This makes logstash::input::udp and ::tcp work the same with respect
to how the $title is used, and repairs line-delimited json log
messages comming in over tcp port 11514.

Change-Id: Idd1bc80d7da0e8f72a009c41d0c9da831fffc795