toolforge-kubeadm: calico upgrade changes
WARNING: The change to liveness probe is not compatible with 3.8.0.

If this is applied to v3.8.0, the calico pods start into a crash loop backoff.
They do work, but that's a lot of unnecessary churn. So don't apply if
v3.8.0 is explicitly set in hiera.

Beyond preparing for the upgrade to 3.14.0 (currently running in toolsbeta),
this allows enabling typha via hiera in the yaml manifest.  Naturally,
this does nothing until applied with kubectl apply by design so we get
a nice rolling upgrade.

Bug: T250863
Change-Id: Icbfcb150f61de8c4862eb96d1fecf3c03b8b03a8