cloud-node-exporter: ignore NFS on the cloud client side
There's a history of NFS + prometheus scraping causing either a thread
leak or an unavailable instance endpoint for the node exporter.

The net effect in Toolforge and Cloud Services is that the monitor ends
up flapping, and we both get alerted and lose metrics.

This seems to be working in tools, where it is set via horizon.

Bug: T252260
Change-Id: I8940966b6f50a0989e7fe24bb70b2c15028174ec