puppet-facts-export-puppetdb: Read localcacert from right section
So this works on puppetmasters that don't use/trust themselves.

Brooke found that on toolsbeta puppetmaster (that trusts the central labs
puppetmaster instead of themselves), this would fail due to PuppetDB presenting
a cert signed by the toolsbeta puppet CA, which this machine controls but does
not actually trust itself.
tools would have this issue too if it had PuppetDB. I can't think of any other
puppetmasters around labs that don't trust themselves.

Bug: T252606
Change-Id: Ib7df84261a0026354c37f31dc68340a7d4324ded