profile::kafka::broker: use openjdk-8 for Buster
What:
Set openjdk-8 when a Kafka broker runs on Debian Buster.

Also add a $java_home in profile::java::java_8 for reference.

Why:
Kafka supports openjdk-11 (the only available on Buster) from 2.1
onward, and we are not there yet. Using openjdk-8 on Buster will
facilitate the migration to the new OS.

Change-Id: I67bdeccd8a532038db382a195fce1483ba76213e