role::archiva: move to profile::java::analytics
What:
Instead of requiring the default jdk, use the specific profile
that Analytics manages.

Why:
Archiva doesn't support (officially) java 11, so we need to put a
conditional on the os version to be able to upgrade to Buster.
Since this is managed by Analytics, let's re-use the existing
profile rather than replicating the code.

Bug: T252767
Change-Id: Ida1fdac80da4bc722bda00400f991eeea4609c99