Resolve duplicate declaration of openjdk-8 on logstash nodes
What:
Avoid duplicate declaration of openjdk-8 on logstash nodes, where
kafka and elasticsearch are running together.

What:
This is not an optimal solution, but it should solve the immediate
problem. Long term, it would be great to have a more standardized
and consistent way of deploying openjdk in our puppet codebase
(like a shared profile), but it will require time and we should
probably reach agreement first.

Change-Id: I3fb00209233653816086a444fc2d8b4bd7a69391