role::druid::analytics::worker: move config to /srv
What:
Move config directives to /srv to standardize druid's
directory location.

Why:
Historically we set a partition for /var/lib/druid, but recently
we moved to a more standard scheme for new hosts, namely using
the /srv partition. This change is part of the procedure to move
all Druid nodes to the new partitioning scheme.

Bug: T252771
Change-Id: Ic794ea760c49268cb5c05895a5132a74974ba189