hiera backends: Comment out special Netbox case
Currently the hiera backend switches to the netbox datasource when an
path in the hierarchy matches the string 'netbox/'.  This means that
when the backend attempts to search in role/common/netbox it is
mistakenly redirected to the netbox backend.

This CR updates the backend so we check for netbox only at the beginning
of the path.

Hosts: O:netbox::frontend

Change-Id: I28b4035626c97a679164d6f417ba34d9b5a5c385