cescout: harden the Postgres installation (improves f3a35978)
This commit improves f3a35978 and adds additional options to harden the
Postgres installation. Notably, it limits read/write access for the
Postgres service to certain directories and restricts the service
capabilities.

(Note about restricting file system access: "ProtectSystem" is already
set to "strict" but that supports read access, and for some directories,
we disable that as well by setting "InacessiblePaths".)

Bug: T247273
Change-Id: I4da3ba12f9f60e4a2245925da4c7d6402b44374e