check_puppet_run_changes: remove staging hosts from this test
The check_puppet_run_changes check runs on the puppetdb servers and
reports all systems that preform a puppet change on every run.
staging/development servers are often in a broken state as people try
new features, software or designs.  Thus this check is often in a
Warning state.

Removing test hosts we would, in theory, mean this check only alerts if
a specific role or profile is broken and not an individual host which by
the nature of being a dev/staging servers could go into a broken puppet
state for days while engineers test new developments.

This CR removes hosts that either:
  * starts with 'labstest'
  * matches the regex: /(?:test|dev)(?:\d{4})?$/

users can run the script with the `--dev` switch to recreate the original
behaviour.

Change-Id: Ie5637e11463904dcf18b4a75659de81059b3aec6