contint: remove obsolete roles
The ci::slave::labs role was for CI permanent Jenkins agents which have
since been phased out. We no more have the role applied anywhere.

ci::slave::lite was a transient role for the new Pipeline. We have since
shifted the dummy jobs it served to the production hosts (was T224069).

Remove rlang support, if we restore testing for that environment, it
will be done through Docker containers.

None of those roles are applied in WMCS.

Bug: T225735
Change-Id: Ie0bd30b21f6df89bb5acb86fc32a4ee1aabe1149