jenkins: master should stick to java 8
When migrating the CI master to Buster and java 11, we had an issue with
the Gearman plugin which is not compatible with java 11. We thus require
the master for CI to stick to java8.

The Jenkins masters are also slaves and those would be fine running
whatever default java version is provided by the OS.

Change the jenkins::slave class to use the default version from the OS,
the agent is run via /usr/bin/java.

The CI Jenkins master should stick to java 8 for now, pin the version in
the systemd service in order to not rely on the alternative selection.

Bug: T224591
Change-Id: I832658c986f9eb28e8c0d347c79e5356f8a79a90