profile::java: one profile to rule them all (openjdk-x versions)
What:
Add a profile that represents the various openjdk-x deployments to
avoid code repetition.

Why:
Due to how openjdk LTS is released, on Buster we have only Java 11
available. This is a problem since several Java-based daemons,
most notably Hadoop and Kafka, do not support Java 11 yet.
The workaround that we have is to deploy a custom backport version
of openjdk-8 on Buster, but it is becoming a copy/paste game
across multiple puppet classes.
This profile also takes care of how to deploy two (or more) versions of Java
at the same time, and how to establish which one is the preferred one
(via alternatives).
By default it just do the most conservative thing, namely deploying Java 11
on Buster and 8 on Stretch.

Credits to ottomata for the final version of the code and Antoine for
the rspec tests.

Signed-off-by: Antoine Musso <hashar@free.fr>
Change-Id: I02262e2638cf2c954ddd5c899fb1b55b6c644eb7