Add testable anycast authdns address and bird stuff
This adds the bird/anycast_healthchecker bits to the authdns-only
servers (it's already present for recdns on dnsbox), and adds the
various bits of hieradata to get it configured and checked in both
cases.

In theory this should result in both profile::dns::auth and
profile::dnsbox (combined rec+auth dns) configuring the new
anycast IP on the loopback, and then monitoring it and advertising
it via bird separately from the similar recdns setup.  We're not
yet adding it to any NS record sets (here or at the registrar)
yet; that process will come after some period of experimentation
and testing.

TODO outside of this patch:
* Router-side support for the chosen subnet and/or the advertised
  IP, at least internally.
* Router-side support for advertising the space publicly (if there
  are local peers for it, but don't try to readvertise it if
  learned from another DC).

Bug: T98006
Change-Id: I8577f003e41154066a80fb875449afee2799feee