varnish: new stap script post_body.stp
Varnish does not support logging request bodies in any way. This script
prints all details of HTTP requests with a body, including the body itself.

The function 'print_http' takes a pointer to a struct http and prints all
headers. It works with both HTTP requests and responses, so it might be
useful in other probes too.

Change-Id: I64c26076b54959edde40de21a5ead64bc9576631