role::druid::analytics::worker: add autoconfig for historical
What:
Set the number of threads and max direct memory size automatically
in the Historical settings if not provided via hiera.

Why:
We are running Druid on different hw, and upstream suggests to set
the number of processing threads for Historical equal to the number
of available cores on the host. As solution, this patch adds autoconfig
for processing threads if not provided in hiera. It also sets max direct
memory size for the JVM accordingly, since a wrong parameter will lead
to the historical refusing to start.

This patch also move the historicals to G1GC, as suggested by upstream.

Bug: T252771
Change-Id: I7d44afd835d3c6ac128e8853431a211986b3e51c