labstore: current setup doesn't allow check_call against exportfs
Since we dropped the annoying bind mounts, running exportfs every 5 min
fails against the DRBD secondary, causing the service to fail as well.

By using just "subprocess.call" it logs a lot of garbage and keeps running.

This needs a bigger fix of adding some smarts to the service to check
for the service IP and not run unless it is there, like maintain-dbusers

Bug: T224582
Change-Id: Ia6137b00abd252dfea7a503b8049b2b68ea7df25