kubeadm and toolforge: rearrange some things to get it working
When kubeadm and toolforge k8s were rearranged in I55eb1869be46bbd5fcb5
Some essential items were shuffled in such a way that they didn't break
Toolforge K8s (anything under /etc/kubernetes/admission changing would
break the control plane quickly because that's mounted in the api), but
it does break bootstrapping anything with the kubeadm module because
those items are excluded from it.  This brings them over to kubeadm.

Also, moving the kubebench materials back to Toolforge because they are
entirely specific to the security needs of TF.  I'm trying to keep this
to functional things only in this patch. Several items are in kubeadm now
that are truly Toolforge specific (calico aliases, metrics setup, likely
the ingress setup, etc.).  Many of those may not be reusable in PAWS
directly.

Bug: T211096
Bug: T251297
Change-Id: Icd0ec8c413da1d36548f9d9294d62512a1f870b0