Add new pool DatabasesCodfw to backup data generated on eqiad
production backups are sent to codfw through replication.

However, the profered method for Databases is to backup eqiad
data and send it to backup2001 (codfw storage) for extra
redundancy over geography locations. This is equivalent, and
the reverse setup to the current Databases pool, whose dumps
and binary backups are generated on codfw and stored on eqiad
storage arrays.

Create a new pool DatabasesCodfw, which uses the backup2001
storage daemon, Databases one.

On a later step, this will be setup to backup from dbprov1001
and dbprov1002.

Also, later, content backups from backup2002 and backup1002
will be sent cross-dc in the same way.

Bug: T79922
Change-Id: I7d7e2873c3aee0e0a6bbc03085aeabd250fbaa2f