alternatives: add a class for the java use case
What:
Add a specific class for update-java-alternatives.

Why:
The class should be useful when dealing with multiple versions
of Java on the same host. The openjdk package ships several binaries
(jstack, jxxx, etc..) that should be aligned with /usr/bin/java.

This is an attempt to introduce the usage of update-java-alternatives
in puppet. After checking some hosts I think that jre/headless/etc..
variants are all going to share the same path, so the simple logic
should work, but please comment in case it is the wrong understanding.

Change-Id: I67119c120fb72c588fa429004221a4bb8dcf57f7