systemd::timer::job: unkludge OnUnitInactiveSec/OnUnitActiveSec
The previous implementation here didn't allow the timer to survive reboots.
This is fixed.

The previous implementation also had a gross kludge of invoking a systemd start
on the service unit to 'get the ball rolling' on OnUnitInactiveSec/OnUnitActiveSec,
as otherwise the service unit has a 'null' last (de)activation timestamp which acts
sort of like NaN, making those trigger conditions meaningless.

The fix is just to add an OnActiveSec interval, if the input interval spec consists
only of OnUnitInactiveSec/OnUnitActiveSec intervals.

Bug: T253374
Change-Id: I3d37ea5031e10a292e5489db3a5f14b10232505a