zuul: add a connection to gerrit-test.wikimedia.org
Attach the production zuul to the test Gerrit instance. The intent is to
fully validate the Gerrit upgrade.  It is essentially a copy paste of
the production Gerrit since the user is the same (jenkins-bot, same
authorized ssh key).

I went with hardcoding the gerrit-test values in the zuul.conf template.
There is no much point in adding all the puppet boilerplate to make it
configurable.

Bug: T253263
Change-Id: I223111b79ea442a80ea133fa6400f4e6857fabbc