profile::tlsproxy::envoy: add max_requests parameter
We want to be able to limit persistent connections that are coming
from a load-balanced client, not just to a load-balanced server, in
some cases.

Specifically, the jobrunners are in that situation: allowing
forever-running persistent connections skews the work of the
round-robin load balancer and risks funneling very expensive requests,
or very busy channels, directly to a single server.

If we interrupt connections after a fixed number of requests,
load-balancing should be effective again.

Change-Id: Id07e7ba5d297ca4605ef5483fdb6fe7a237f38c6