vcl: apply mobileaction/useformat ttl cap to cacheable responses
An ancient comment dating back to de0592289 says "front page purges
aren't going to come through for this url, so don't cache it for long",
indicating that this VCL code was meant as a TTL cap. To avoid the risk
of caching 50x or other uncacheable responses, ensure we apply the cap
only to responses with TTL higher than one minute.

Bug: T247783
Change-Id: I585fe9fa4c36dd631d8b74c822090ff771edada6