dynamicproxy: Short-circuit urlproxy lookups against canonical_domain
Urlproxy should only do path based backend lookups for grid engine
hosted webservices when the inbound hostname associated with the request
is "tools.wmflabs.org" (or the deployment specific equivalent). To
ensure this, we introduce a conditional check around subdomain lookups
to ensure that the hostname ends with the expected canonical domain. If
this test passes, we will not cascade into path based lookups. Instead
we either find a match in the active grid services and set the backend
appropriately, or we exit without setting the backend and let control
pass on to the Kubernetes cluster.

A minor optimization is also made to the nginx config to remove an `if`
because <https://www.nginx.com/resources/wiki/start/topics/depth/ifisevil/>.

Bug: T253816
Change-Id: I00e3270d5e09ab0dfdb97a820bc4a466a37556df