Gemfile: fix pry / pry-byebug incompatibility
We still support ruby 2.3 that came from Stretch and thus require to pin
pry-byebug to 3.7.x. It in turn as a loose dependency on pry but it is
latest 0.13.0 release is a breaking change.

pry-byebug 3.9.0 brings compatibilty with pry 0.13.0 and also strengthen
the version requirements to avoid bringing potentially incompatible
minor release. But we can not use pry-byebug 3.9.0 since we would loose
ruby 2.3 support.

Instead pin pry to the previous minor release (0.12.2). That unbreaks
running Beaker for me on Buster / ruby 2.5.

Bug: T253635
Change-Id: I28664f77d2a6d7c1b0e8e7ea6713d3dc4bcb64ca