ATS: add planet.wikimedia.org to also map to its backend
In addition to the language versions (en.planet.wikimedia.org,
de.planet.wikimedia.org etc) also add just "planet" without a
language and map it to the same backend.

To be able to add an index page or redirect and avoid getting
the "Our servers are currently under maintenance" error page.

Change-Id: I21e21617cf03a45f4796bab639e5a60eea28313e