kubeadm: wmcs-k8s-node-upgrade: don't skip the pause stage
Don't skip the pause stage unless it was requested by a cmdline switch.

This prevents a disaster situation in which a node was skipped in a later stage
(upgrade), which means the node is disabled. In such situation, skipping all
nodes in such stage could mean all the workers end depooled at the same time.

Bug: T250867
Change-Id: Id313a1c1602089d161a73becf3cf4b15fa2ae08a
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>