Consolidate query_service profile duplication
A followup patch will be adding a third configuration of
blazegraph for sdoc. Prepare by sharing as much as possible between
the two current installations, avoiding more duplication when the third
is added.

Changes profile::query_service::blazegraph from a class into a define,
and adds two new profiles, one for each installation, that invoke it.

Change-Id: I464c4125fa43d9e3ca3d9fbe48fe99f4e2a744a7