transfer.py: Use xtrabackup on path now that all packages are 10.1.43+
In previous version of the package, there was no xtrabackup alias.
Make the application more universal by using xtrabackup on path, so
it can be used for the wmf-mariadb101 wmf-mariadb104 or other
packages.

This was hardcoded to /opt/wmfmariadb101/bin patch before as those
packages didn't use update-alternatives to select the prefered
xtrabackup binary.

Make also mbstream version-independent by setting it on path, until
all packages have been upgraded to the later versions where mbstream's
path is part of update-alternatives.

Change-Id: I37eddf1729e19d3cc23babb72247cf15f425fa8b