mediawiki: Add api.wikimedia.org
api.wikimedia.org is the new wiki that will act as documentation for
the API Gateway. it has already been configured on MediaWiki, so now
we want to make it accessible to the public.

This domain is currently accessed via Envoy running in k8s rather than
via the traditional path, but we still need appservers to serve it as
normal.

Bug: T246945
Change-Id: I56a6d70bf47a2229c8f56641eb3661baeab47cb3