peek: Reenable cron with correct params
First of each week cron: 0 0 * * MON
First of each month cron: 0 0 1 * *

Note to self, cron anatomy params not
specified default to '*' and not 0.

Change redirect to send stdout to /dev/null
with stderror respecting the .forward
specification of security-team@ to not
flood root@ on issue.

Created a task to convert to systemd timer which
will take a bit more figuring than I have time
for now, and I want to talk to the security team
about monitoring strategy, etc.

Bug: T254127
Bug: T251784
Bug: T254368
Change-Id: Ibb004376aebbd2a72f3d31a0230a8cb5d0f98396