dumps-distribution: don't monitor systemd directly for paging
As is, we have systemd timers and services that alert 2 different teams.
Monitoring systemd as a whole will always page WMCS because the host
pages WMCS. It isn't trivial to set a different "team" (we have an email-
only "team" for wmcs) to monitor systemd, so we'd like to simply *not*
monitor that directly, which is trivial.

This change should only affect labstore1006/7.

Change-Id: I72a9f1f8c67568e969998f5256529dd7ab44d80b