puppetmaster: update puppet-merge
puppet merge currently will halt merging the production repo if a user
declines to merge the labs private repo.  This CR updates the exits code
used when a users cancels a merge to 99 so that it will be treated in
the same manner as if there was nothing to merge.

We also continue the script on merge errors however we only preform the
remote puppet-merge if the local puppet-merge exited with 0

Further the private repo doesn't need to exist on all worker nodes, only
the two frontend servers.  This CR also updates the script so the puppet
private repo is only synced with the puppet masters.

In theory the puppetmasters don't need the private repo at all, they just
need a way to have a user validate the diff between the sha1 in
 /srv/config-master/labsprivate-sha1.txt
 against the current master and if safe to accept update this file with
 master sha1.  however this is a much more involved rewrite

Hosts: C:puppetmaster

Bug: T251104
Change-Id: I5d45540e51eacb9f403bc379f89d5635012ce121