beta: Allow using docker volumes
Why:
Horizon has a 64KB limit on the hiera that can be passed and changeprop
has a configuration that is considerably larger. The alternate approach
of using puppet.git hieradata/cloud was considered, however that approach
greatly hinders CPT, as they have to bug SRE a +2 every time they want
to change a setting in changeprop.

What:
Allow specifying a parameter named volume that instructs docker run to
use a named docker volume instead.

The docker volume should be named whatever $title is and is expected to
contain whatever puppet+horizon hiera would populate
/etc/$title/config.yaml as.

Bug: T251176
Change-Id: Ifeb26c20a23d27693de6029bcc71909135d2311d