dnsdist: add a parameter for setting addDOHLocal's base URL
For the addDOHLocal function, the "urls" parameter is defined as "a base
URL, or a list of base URLs, to accept queries on." We set it to
"/dns-query" and update the dnsdist.conf template.

This change also helps fix a bug with 3a9c7f79, as the arguments to
addDOHLocal are now passed in the correct order.

Bug: T252132
Change-Id: I53e3e0a807eb128969bc364b12faff134d3f8a65