centrallog: split syslogs into host directories
Currently centrallog outputs syslogs from all hosts into a common set of
flat files. These files become large, and searching through them for
something speicfic can be time consuming.

To help with this, let's split syslogs off into hostname directories like this:

  /srv/syslog/$hostname/syslog.log

Along with that, let's update logrotate to rotate and compress files in the
hostname directory so they are easier/faster to find in the future. like this:

  /srv/syslog/$hostname/syslog.log-$datestamp.gz

The existing programname splits for apache2.log, swift.log and the syslog.log
catch-all are preserved (within the hostname directories).

Change-Id: I736e48702680f6a5c1bc886e8c6031911cac4096