profile::etcd::tlsproxy: add testserver ACLs
The conftool "testserver" cluster is used for various things,
including scap dsh lists, but it doesn't have an LVS endpoint because
we use an L7 load-balancer there (we need to direct requests to the
right backend based on a header).

So it was not accounted for in our list of ACLs. Add it manually to
the list. Also, add all services and not just the ones in production -
I can imagine someone wanting to depool a server before the service is
actually in production.

Change-Id: I84adc9fc6c0bbdf71f68225412428480479a7bec