mariadb: Add snapshotting capabilities to content backup hosts
Add the mariadb::backup::snapshot profile to backup1002 and
backup2002, installing the wmf mariadb package and setting up
the snapshot directory hierarchy, so they are ready to receive
and process backups, even if no snapshots are regularly
scheduled there.

This will allow an emergency snapshot/temporary storage in a host
with enough free space, if necessary.

The immediate usage would be to test snapshotting in a buster/
mariadb 10.4 workflow before updating the production dbprov hosts
to it.

Bug: T79922
Bug: T250666
Change-Id: I0dd3c2f312032268253982918192ca34e79ce3a0