wikistats (cloud): update query for XML dump of Wikipedia table
Apparently there was an issue with escaping since the query
got cut off (when watching /var/log/syslog for action triggered
by the cron job).

The same result can be achieved by checking whether total IS NOT NULL,
avoiding the need to escape anything.

Why this issue showed up now and not earlier is unknown.

Bug: T254214
Change-Id: Idc2d4bc678a9c1c665384f9a9eadff3297b9bb61