phabricator: change sender address of community_metrics mail
When testing project_changes.sh and community_metrics.sh
after I0f02a02877800, I noticed that the latter did not send
email anymore.

The reason being sender verification by the mail servers and
the address/alias communitymetrics@ being gone.

I am not sure if that existed until recently on the Google side
and was now dropped (I asked OIT with a Zendesk ticket) or if
it started breaking now due to sender verification becoming
more strict.

Change it to also use the personal aklapper@ address as sender,
like the project_changes script already does.

Though if we want to avoid that and use something generic we
would have to ask OIT to create one or maintain it in exim aliases
on our side, which we would like to avoid too.

Change-Id: I6384feb32d76af4a094b66934cad20e1f0aadf10