webperf: Remove XHGui dependency on MongoDB
This moves to installing a Debianized version of XHGui, which is based
on a newer release and supports storing profile data in a SQL database
instead of MongoDB.

Before merging, this will need the database to be set up and the
secrets added to puppet-private.  I'll also need some help uploading
the packages from performance/debs.  (I've verified they build cleanly
and work locally using `gbp buildpackage`.)

Bug: T180761
Change-Id: Icc7378565d49cf126786f6e2bbd42c4695224086