run-puppet-agent: add new flag --unless-version SUBSTR
run-puppet-agent --unless-version SUBSTR will run puppet only if
the current configuration version string (e.g. "(2d1d2c1543) CDanis -
expand phpfpm status text exporter to all appservers") does not match
the given SUBSTR, which is best given as e.g. CDanis or 2d1d2c1543.

This is useful for when you've merged a change and manually run puppet on
some of the fleet for limited testing, after you disabled puppet elsewhere,
and then after tests you want to quickly deploy it on the remaining servers.

Not re-using in the 97-last-puppet-run MOTD script, as this approach is
too slow for use on every login.

Also fix a hilarious bug created during manual testing of this patch itself:
https://phabricator.wikimedia.org/P11435

Change-Id: I1deeb25843ad6a17028c19dd5774aebc81e4ffc9