memcached: allow more tunables to avoid implicit settings
Memcached's usage seems expanding a lot, more systems are using it
and we should make defaults more explicit.
For example, we are setting in a lot of places the min slab size of 5,
instead of the default (48), that is/was a specific setting for mediawiki's
object cache. This change aims to be a no-op, all the memcached settings
to be reviewed have been annotated with a TODO.

Bug: T252391
Change-Id: I6b8104e3605ad56cc3e29b85153933b3e21021c1