wikidough: update profile to use a single recursor
For the initial version of wikidough, the dnsdist instance will use a
single recursive resolver instead of multiple ones. This commit updates
the wikidough profile and the dnsdist class to reflect that change.

While we could have continued using the existing multiple resolver
configuration (fan-out), the reason behind this commit is to correctly
reflect the current design of wikidough.

Bug: T252132
Change-Id: I84a650611f4bdf33ebf058022df7e12978e987bd