Gerrit: Enable crons on Gerrit replicas
On gerrit2001 (our only Gerrit replica), logs pile up and do not get
cleaned up. So we turn on Gerrit crons to get them cleaned up
automatically.

The only other thing the crons currently do is to create a list of
gerrit extensions every 15 minutes. Creating that list is not
explicitly required, but it is cheap and does not hurt.

Change-Id: Ie24ead1f2388dd42604e8b321b39945923a7eabd