dnsrecursor: make forward-zones and edns-subnet-whitelist optional
The forward-zones and edns-subnet-whitelist settings for pdns-recursor's
configuration are currently hardcoded as the class was intended for use
as an internal resolver. To use the dnsrecursor class for setting up
Wikidough's recursive resolver (external use), we should avoid setting
these options.

This commit updates the dnsrecursor class and the recursor.conf template
by adding boolean parameters for both these options, setting them to
true to maintain compatibility with existing use, but allowing them to
not be set if required, such as for its use in Wikidough.

Bug: T252132
Change-Id: I2dfc2a7b11499946892217bb980ea83b77804162