gerrit: add parameter for db_name, let gerrit1002 use test db
Previously, the gerrit test server, gerrit1002, had access to
the production DB but using a separate readonly user.

Now, a separate database has been created as a clone of the
production database, so the test server can have read-write access.

To make it possible to change this without having to disable puppet
entirely, add a parameter that allows changing the database name
in Hiera, along with existing db_user and db_pass.

The password for the test database also has to be added in the private
repo for this to work.

Bug: T254516
Change-Id: Ie10d89a65432242ce22febb35ebed2e157175785