purged: make Kafka cluster name configurable
Add a new hiera attribute, profile::cache::purge::kafka_cluster_name, to
be used to lookup the Kafka configuration with kafka_config().

The value should be set in production to the nearest main DC and
defaults to 'main-eqiad'. For DCs closer to codfw than to eqiad, namely
ulsfo and codfw itself, we set the value in hiera to 'main-codfw'. Set
the value to 'main-deployment-prep' for deployment-prep.

Hosts: cp2027.codfw.wmnet, cp3050.esams.wmnet, cp4023.ulsfo.wmnet

Bug: T254844
Bug: T133821
Bug: T250781
Change-Id: Ib8fb3d56a02d42beecff45c9bd3d31d0afa9f336