wmcs: kubeadm: introduce different haproxy port frontend/backend
Up until now, we had a simplified template because we didn't need otherwise in toolforge. Now with
the PAWS cluster, the haproxy is doing SSL termination too (future patch) and therefore we need
to set up more stuff in the haproxy config.

This patch introduces a hiera key that defines frontend/backend ports.

Bug: T195217
Change-Id: I231a5fc5f253d1f23eb522bbc12c330d47a9ccb4
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>