labstore: fix the failover process
The failover process with DRBD now actually works since the abolishing of
the bind mounts. It does still have 2 issues, however.

1. Running nfs-exportd before starting the NFS server is a serious delay
at getting the server working, doesn't work correctly anyway and is largely
pointless because nfs-exportd is already running on the server. Ensuring
that exportfs is run after the server is up is good, and the nfs-exportd
run does that as well as a quick update on the files.

2. The umount needs the -f option. Leaving it off just requires a manual
step. It works fine once that's there.

Bug: T224582
Bug: T169570
Change-Id: Ic4bc902578c3d985f0dea5a058505368fe99cae2