archiva: assign archiva-new.wikimedia.org to archiva1002
This is part of the migration of archiva.wikimedia.org's backend
(archvia1001) to Debian Buster. The idea is to add another domain
that people can test, archiva-new.wikimedia.org, before making the
switch. This procedure has already been tested in the past with
meitnerium -> archiva1001.

The delicate part is the let's encrypt certificate, that uses
the ACME protocol to establish who is the authority for a TLS
certificate (who in the case is archiva1001 or archiva1002).
To avoid any risk of takeover before the new host is tested,
we created archiva-new as CNAME for archiva1002.

Bug: T252767
Change-Id: I7bd37b71adc9668233ac4df1baff308d999fc52a