wmcs: kubeadm: haproxy: introduce support for TLS termination
This patch introduces support TLS termination in haproxy. For now, the haproxy server will contact
the backends without doing TLS.

The TLS config is optional, and will only be enabled if required via hiera.

We plan to use TLS termination in haproxy for the PAWS kubernetes cluster, however Toolforge is
doing it in the front-proxy, before haproxy.

Bug: T195217
Change-Id: Ic6deb5f955a2a5361f929deffafbe26e6bd1c0a7
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>