cloud nfs: allow opt-in soft mounting wherever folks want to try it
Soft mounts are better for server stability and worse for data safety.
Where data is changing often and corruption is a serious concern, keep
the default. If the server mostly keeps critical data on the instance's
disk, it might be safer to use soft mounts.

Soft mounting /home seems a bit cavalier. However, if that's how you roll
we will be able to elect to do it if desired. Because of how the maps
volume has been mounted, it's all treated as a home dir until it gets
a different mount setup.

Bug: T127559
Change-Id: Ibb0f6014462f0cab060579a7995c9b708cafde82