services_proxy: Add keepalive, retries
Why:
We want the same protections for termbox as for the various eventgate
services as we are probably noticing a race condition of the mw envoy
trying to route a request through a connection that is being in the
meantime closed by the termbox sidecar envoy

What:
Add a keepalive of 4.5s and a retry on 5xx of 1. Amend the timeout as
well to be larger the keepalive. Align it with eventgate-main

Bug: T255410
Change-Id: Ib7bfacf88eef436ed0f34a16e9d39eb0c053a44f