ATS: stop caching noc.wm.org responses
The origin server for noc.wikimedia.org was changed from
mwmaint1002.eqiad.wmnet to mwmaint.discovery.wmnet in 03d659ba. The
corresponding caching rule was however not updated, with the result that
we are currently caching noc.wm.org responses at the ats-be layer, and
we shouldn't.

Use mwmaint.discovery.wmnet in the rule.

Bug: T255368
Bug: T210411
Change-Id: Ib606697edc77e7df0994e498fd8f35de3ad2545d